@@ -1,7 +1,7 @@
-project(SimpleLogger)
+cmake_minimum_required(VERSION 3.10)
+cmake_policy(VERSION 3.10)
 
-cmake_minimum_required(VERSION 2.8.12)
-cmake_policy(VERSION 2.8.12)
+project(SimpleLogger)
 
 option(BUILD_TESTS "Build unit tests" ON)
 
@@ -12,16 +12,11 @@ if(NOT CMAKE_BUILD_TYPE)
 MinSizeRel." FORCE)
 endif()
 
-set(CMAKE_CXX_STANDARD 11)
+set(CMAKE_CXX_STANDARD 17)
 
 if(CMAKE_COMPILER_IS_GNUCXX OR MINGW OR ${CMAKE_CXX_COMPILER_ID} STREQUAL "Clang")
     set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -Wall -Wextra -Wpedantic")
     set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -fvisibility-inlines-hidden")
-
-    # CMAKE_CXX_STANDARD supported only by versions >= 3.1
-    if (CMAKE_VERSION VERSION_LESS "3.1")
-        set (CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -std=gnu++11")
-    endif ()
 endif()
 
 set(LIBRARY_NAME "SimpleLogger")
