@@ -27,338 +27,398 @@
 #include <chrono>
 #include <ctime>
 #include <fstream>
+#include <iomanip>
 #include <iostream>
 #include <map>
 #include <mutex>
 #include <stdexcept>
 
-#ifdef Q_OS_ANDROID
-#include <QDebug>
-#endif
-
 namespace juzzlin {
 
-class Logger::Impl
+class SimpleLogger::Impl
 {
 public:
-
     Impl();
+
+    Impl(const std::string & tag);
 
     ~Impl();
 
-    std::ostringstream & trace();
+    std::ostringstream & traceStream();
 
-    std::ostringstream & debug();
+    std::ostringstream & debugStream();
 
-    std::ostringstream & info();
+    std::ostringstream & infoStream();
 
-    std::ostringstream & warning();
+    std::ostringstream & warningStream();
 
-    std::ostringstream & error();
+    std::ostringstream & errorStream();
 
-    std::ostringstream & fatal();
+    std::ostringstream & fatalStream();
 
     static void enableEchoMode(bool enable);
 
-    static void setLevelSymbol(Logger::Level level, std::string symbol);
+    static void setLevelSymbol(SimpleLogger::Level level, std::string symbol);
 
-    static void setLoggingLevel(Logger::Level level);
+    static void setLoggingLevel(SimpleLogger::Level level);
 
-    static void setTimestampMode(Logger::TimestampMode timestampMode, std::string separator);
+    static void setCustomTimestampFormat(std::string format);
+
+    static void setTimestampMode(SimpleLogger::TimestampMode timestampMode);
+
+    static void setTimestampSeparator(std::string separator);
 
     static void setStream(Level level, std::ostream & stream);
 
-    static void init(std::string filename, bool append);
+    static void initialize(std::string filename, bool append);
 
     void flush();
 
-    std::ostringstream & getStream(Logger::Level level);
-
-    void prefixTimestamp();
+    std::ostringstream & prepareStreamForLoggingLevel(SimpleLogger::Level level);
 
 private:
+    std::string currentDateTime(std::chrono::time_point<std::chrono::system_clock> now, const std::string & dateTimeFormat) const;
+
+    void flushFileIfOpen();
+
+    void flushEchoIfEnabled();
+
+    void prefixWithLevelAndTag(SimpleLogger::Level level);
+
+    void prefixWithTimestamp();
+
+    bool shouldFlush() const;
 
     static bool m_echoMode;
 
-    static Logger::Level m_level;
+    static SimpleLogger::Level m_level;
 
-    static Logger::TimestampMode m_timestampMode;
+    static SimpleLogger::TimestampMode m_timestampMode;
 
     static std::string m_timestampSeparator;
 
-    static std::ofstream m_fout;
+    static std::string m_customTimestampFormat;
 
-    using SymbolMap = std::map<Logger::Level, std::string>;
+    static std::ofstream m_fileStream;
+
+    using SymbolMap = std::map<SimpleLogger::Level, std::string>;
     static SymbolMap m_symbols;
 
-    using StreamMap = std::map<Logger::Level, std::ostream *>;
+    using StreamMap = std::map<SimpleLogger::Level, std::ostream *>;
     static StreamMap m_streams;
 
     static std::recursive_mutex m_mutex;
 
-    Logger::Level m_activeLevel = Logger::Level::Info;
+    SimpleLogger::Level m_activeLevel = SimpleLogger::Level::Info;
 
     std::lock_guard<std::recursive_mutex> m_lock;
 
-    std::ostringstream m_oss;
+    std::string m_tag;
+
+    std::ostringstream m_message;
 };
 
-bool Logger::Impl::m_echoMode = true;
+bool SimpleLogger::Impl::m_echoMode = true;
 
-Logger::Level Logger::Impl::m_level = Logger::Level::Info;
+SimpleLogger::Level SimpleLogger::Impl::m_level = SimpleLogger::Level::Info;
 
-Logger::TimestampMode Logger::Impl::m_timestampMode = Logger::TimestampMode::DateTime;
+SimpleLogger::TimestampMode SimpleLogger::Impl::m_timestampMode = SimpleLogger::TimestampMode::DateTime;
 
-std::string Logger::Impl::m_timestampSeparator = ": ";
+std::string SimpleLogger::Impl::m_timestampSeparator = ": ";
 
-std::ofstream Logger::Impl::m_fout;
+std::string SimpleLogger::Impl::m_customTimestampFormat;
+
+std::ofstream SimpleLogger::Impl::m_fileStream;
 
 // Default level symbols
-Logger::Impl::SymbolMap Logger::Impl::m_symbols = {
-    {Logger::Level::Trace,   "T:"},
-    {Logger::Level::Debug,   "D:"},
-    {Logger::Level::Info,    "I:"},
-    {Logger::Level::Warning, "W:"},
-    {Logger::Level::Error,   "E:"},
-    {Logger::Level::Fatal,   "F:"}
+SimpleLogger::Impl::SymbolMap SimpleLogger::Impl::m_symbols = {
+    { SimpleLogger::Level::Trace, "T:" },
+    { SimpleLogger::Level::Debug, "D:" },
+    { SimpleLogger::Level::Info, "I:" },
+    { SimpleLogger::Level::Warning, "W:" },
+    { SimpleLogger::Level::Error, "E:" },
+    { SimpleLogger::Level::Fatal, "F:" }
 };
 
 // Default streams
-Logger::Impl::StreamMap Logger::Impl::m_streams = {
-    {Logger::Level::Trace,   &std::cout},
-    {Logger::Level::Debug,   &std::cout},
-    {Logger::Level::Info,    &std::cout},
-    {Logger::Level::Warning, &std::cerr},
-    {Logger::Level::Error,   &std::cerr},
-    {Logger::Level::Fatal,   &std::cerr}
+SimpleLogger::Impl::StreamMap SimpleLogger::Impl::m_streams = {
+    { SimpleLogger::Level::Trace, &std::cout },
+    { SimpleLogger::Level::Debug, &std::cout },
+    { SimpleLogger::Level::Info, &std::cout },
+    { SimpleLogger::Level::Warning, &std::cerr },
+    { SimpleLogger::Level::Error, &std::cerr },
+    { SimpleLogger::Level::Fatal, &std::cerr }
 };
 
-std::recursive_mutex Logger::Impl::m_mutex;
+std::recursive_mutex SimpleLogger::Impl::m_mutex;
 
-Logger::Impl::Impl()
-  : m_lock(Logger::Impl::m_mutex)
+SimpleLogger::Impl::Impl()
+  : m_lock(m_mutex)
 {
 }
 
-Logger::Impl::~Impl()
+SimpleLogger::Impl::Impl(const std::string & tag)
+  : m_lock(m_mutex)
+  , m_tag(tag)
+{
+}
+
+SimpleLogger::Impl::~Impl()
 {
     flush();
 }
 
-void Logger::Impl::enableEchoMode(bool enable)
+void SimpleLogger::Impl::enableEchoMode(bool enable)
 {
-    Impl::m_echoMode = enable;
+    m_echoMode = enable;
 }
 
-std::ostringstream & Logger::Impl::getStream(Logger::Level level)
+std::ostringstream & SimpleLogger::Impl::prepareStreamForLoggingLevel(SimpleLogger::Level level)
 {
     m_activeLevel = level;
-    Impl::prefixTimestamp();
-    m_oss << Impl::m_symbols[level] << " ";
-    return m_oss;
+    prefixWithTimestamp();
+    prefixWithLevelAndTag(level);
+    return m_message;
 }
 
-void Logger::Impl::setLevelSymbol(Level level, std::string symbol)
+void SimpleLogger::Impl::setLevelSymbol(Level level, std::string symbol)
 {
-    Impl::m_symbols[level] = symbol;
+    m_symbols[level] = symbol;
 }
 
-void Logger::Impl::setLoggingLevel(Logger::Level level)
+void SimpleLogger::Impl::setLoggingLevel(SimpleLogger::Level level)
 {
-    Impl::m_level = level;
+    m_level = level;
 }
 
-void Logger::Impl::setTimestampMode(TimestampMode timestampMode, std::string separator)
+void SimpleLogger::Impl::setCustomTimestampFormat(std::string customTimestampFormat)
 {
-    Impl::m_timestampMode = timestampMode;
-    Impl::m_timestampSeparator = separator;
+    m_customTimestampFormat = customTimestampFormat;
 }
 
-void Logger::Impl::prefixTimestamp()
+void SimpleLogger::Impl::setTimestampMode(TimestampMode timestampMode)
 {
-    std::string timeStr;
+    m_timestampMode = timestampMode;
+}
+
+void SimpleLogger::Impl::setTimestampSeparator(std::string separator)
+{
+    m_timestampSeparator = separator;
+}
+
+std::string SimpleLogger::Impl::currentDateTime(std::chrono::time_point<std::chrono::system_clock> now, const std::string & dateTimeFormat) const
+{
+    std::ostringstream oss;
+    const auto rawTime = std::chrono::system_clock::to_time_t(now);
+    oss << std::put_time(std::localtime(&rawTime), dateTimeFormat.c_str());
+
+    return oss.str();
+}
+
+void SimpleLogger::Impl::prefixWithLevelAndTag(SimpleLogger::Level level)
+{
+    m_message << m_symbols[level] << (!m_tag.empty() ? " " + m_tag + ":" : "") << " ";
+}
+
+void SimpleLogger::Impl::prefixWithTimestamp()
+{
+    std::string timestamp;
 
     using std::chrono::duration_cast;
     using std::chrono::system_clock;
 
-    switch (Impl::m_timestampMode)
-    {
-    case Logger::TimestampMode::None:
+    switch (m_timestampMode) {
+    case SimpleLogger::TimestampMode::None:
         break;
-    case Logger::TimestampMode::DateTime:
-    {
-        time_t rawTime;
-        time(&rawTime);
-        timeStr = ctime(&rawTime);
-        timeStr.erase(timeStr.length() - 1);
-    }
+    case SimpleLogger::TimestampMode::DateTime: {
+        timestamp = currentDateTime(system_clock::now(), "%a %b %e %H:%M:%S %Y");
+    } break;
+    case SimpleLogger::TimestampMode::ISODateTime: {
+        timestamp = currentDateTime(system_clock::now(), "%Y-%m-%dT%H:%M:%S");
+    } break;
+    case SimpleLogger::TimestampMode::EpochSeconds:
+        timestamp = std::to_string(duration_cast<std::chrono::seconds>(system_clock::now().time_since_epoch()).count());
         break;
-    case Logger::TimestampMode::EpochSeconds:
-        timeStr = std::to_string(duration_cast<std::chrono::seconds>(system_clock::now().time_since_epoch()).count());
+    case SimpleLogger::TimestampMode::EpochMilliseconds:
+        using std::chrono::duration_cast;
+        timestamp = std::to_string(duration_cast<std::chrono::milliseconds>(system_clock::now().time_since_epoch()).count());
         break;
-    case Logger::TimestampMode::EpochMilliseconds:
-        timeStr = std::to_string(duration_cast<std::chrono::milliseconds>(system_clock::now().time_since_epoch()).count());
+    case SimpleLogger::TimestampMode::EpochMicroseconds:
+        using std::chrono::duration_cast;
+        timestamp = std::to_string(duration_cast<std::chrono::microseconds>(system_clock::now().time_since_epoch()).count());
         break;
-    case Logger::TimestampMode::EpochMicroseconds:
-        timeStr = std::to_string(duration_cast<std::chrono::microseconds>(system_clock::now().time_since_epoch()).count());
+    case SimpleLogger::TimestampMode::Custom:
+        timestamp = currentDateTime(system_clock::now(), m_customTimestampFormat);
         break;
     }
 
-    if (!timeStr.empty())
-    {
-        m_oss << timeStr << m_timestampSeparator;
+    if (!timestamp.empty()) {
+        m_message << timestamp << m_timestampSeparator;
     }
 }
 
-void Logger::Impl::flush()
+bool SimpleLogger::Impl::shouldFlush() const
 {
-    if (m_activeLevel < m_level)
-    {
-        return;
-    }
+    return m_activeLevel >= m_level && !m_message.str().empty();
+}
 
-    if (!m_oss.str().size())
-    {
-        return;
+void SimpleLogger::Impl::flushFileIfOpen()
+{
+    if (m_fileStream.is_open()) {
+        m_fileStream << m_message.str() << std::endl;
+        m_fileStream.flush();
     }
+}
 
-    if (Impl::m_fout.is_open())
-    {
-        Impl::m_fout << m_oss.str() << std::endl;
-        Impl::m_fout.flush();
-    }
-
-    if (Impl::m_echoMode)
-    {
-#ifdef Q_OS_ANDROID
-        qDebug() << m_oss.str().c_str();
-#else
-        auto stream = Impl::m_streams[m_activeLevel];
-        if (stream) {
-            *stream << m_oss.str() << std::endl;
+void SimpleLogger::Impl::flushEchoIfEnabled()
+{
+    if (m_echoMode) {
+        if (auto && stream = m_streams[m_activeLevel]; stream) {
+            *stream << m_message.str() << std::endl;
             stream->flush();
         }
-#endif
     }
 }
 
-void Logger::Impl::init(std::string filename, bool append)
+void SimpleLogger::Impl::flush()
 {
-    if (!filename.empty())
-    {
-        Impl::m_fout.open(filename, append ? std::ofstream::out | std::ofstream::app : std::ofstream::out);
-        if (!Impl::m_fout.is_open())
-        {
+    if (shouldFlush()) {
+        flushFileIfOpen();
+        flushEchoIfEnabled();
+    }
+}
+
+void SimpleLogger::Impl::initialize(std::string filename, bool append)
+{
+    if (!filename.empty()) {
+        m_fileStream.open(filename, append ? std::ofstream::out | std::ofstream::app : std::ofstream::out);
+        if (!m_fileStream.is_open()) {
             throw std::runtime_error("ERROR!!: Couldn't open '" + filename + "' for write.\n");
         }
     }
 }
 
-std::ostringstream & Logger::Impl::trace()
+std::ostringstream & SimpleLogger::Impl::traceStream()
 {
-    return getStream(Logger::Level::Trace);
+    return prepareStreamForLoggingLevel(SimpleLogger::Level::Trace);
 }
 
-std::ostringstream & Logger::Impl::debug()
+std::ostringstream & SimpleLogger::Impl::debugStream()
 {
-    return getStream(Logger::Level::Debug);
+    return prepareStreamForLoggingLevel(SimpleLogger::Level::Debug);
 }
 
-std::ostringstream & Logger::Impl::info()
+std::ostringstream & SimpleLogger::Impl::infoStream()
 {
-    return getStream(Logger::Level::Info);
+    return prepareStreamForLoggingLevel(SimpleLogger::Level::Info);
 }
 
-std::ostringstream & Logger::Impl::warning()
+std::ostringstream & SimpleLogger::Impl::warningStream()
 {
-    return getStream(Logger::Level::Warning);
+    return prepareStreamForLoggingLevel(SimpleLogger::Level::Warning);
 }
 
-std::ostringstream & Logger::Impl::error()
+std::ostringstream & SimpleLogger::Impl::errorStream()
 {
-    return getStream(Logger::Level::Error);
+    return prepareStreamForLoggingLevel(SimpleLogger::Level::Error);
 }
 
-std::ostringstream & Logger::Impl::fatal()
+std::ostringstream & SimpleLogger::Impl::fatalStream()
 {
-    return getStream(Logger::Level::Fatal);
+    return prepareStreamForLoggingLevel(SimpleLogger::Level::Fatal);
 }
 
-void Logger::Impl::setStream(Level level, std::ostream & stream)
+void SimpleLogger::Impl::setStream(Level level, std::ostream & stream)
 {
-    Logger::Impl::m_streams[level] = &stream;
+    m_streams[level] = &stream;
 }
 
-Logger::Logger()
-    : m_impl(new Logger::Impl)
+SimpleLogger::SimpleLogger()
+  : m_impl(std::make_unique<SimpleLogger::Impl>())
 {
 }
 
-void Logger::init(std::string filename, bool append)
+SimpleLogger::SimpleLogger(const std::string & tag)
+  : m_impl(std::make_unique<SimpleLogger::Impl>(tag))
 {
-    Impl::init(filename, append);
 }
 
-void Logger::enableEchoMode(bool enable)
+void SimpleLogger::initialize(std::string filename, bool append)
+{
+    Impl::initialize(filename, append);
+}
+
+void SimpleLogger::enableEchoMode(bool enable)
 {
     Impl::enableEchoMode(enable);
 }
 
-void Logger::setLoggingLevel(Level level)
+void SimpleLogger::setLoggingLevel(Level level)
 {
     Impl::setLoggingLevel(level);
 }
 
-void Logger::setLevelSymbol(Level level, std::string symbol)
+void SimpleLogger::setLevelSymbol(Level level, std::string symbol)
 {
     Impl::setLevelSymbol(level, symbol);
 }
 
-void Logger::setTimestampMode(TimestampMode timestampMode, std::string separator)
+void SimpleLogger::setTimestampMode(TimestampMode timestampMode)
 {
-    Impl::setTimestampMode(timestampMode, separator);
+    Impl::setTimestampMode(timestampMode);
 }
 
-void Logger::setStream(Level level, std::ostream & stream)
+void SimpleLogger::setCustomTimestampFormat(std::string customTimestampFormat)
+{
+    Impl::setTimestampMode(TimestampMode::Custom);
+    Impl::setCustomTimestampFormat(customTimestampFormat);
+}
+
+void SimpleLogger::setTimestampSeparator(std::string timestampSeparator)
+{
+    Impl::setTimestampSeparator(timestampSeparator);
+}
+
+void SimpleLogger::setStream(Level level, std::ostream & stream)
 {
     Impl::setStream(level, stream);
 }
 
-std::ostringstream & Logger::trace()
+std::ostringstream & SimpleLogger::trace()
 {
-    return m_impl->trace();
+    return m_impl->traceStream();
 }
 
-std::ostringstream & Logger::debug()
+std::ostringstream & SimpleLogger::debug()
 {
-    return m_impl->debug();
+    return m_impl->debugStream();
 }
 
-std::ostringstream & Logger::info()
+std::ostringstream & SimpleLogger::info()
 {
-    return m_impl->info();
+    return m_impl->infoStream();
 }
 
-std::ostringstream & Logger::warning()
+std::ostringstream & SimpleLogger::warning()
 {
-    return m_impl->warning();
+    return m_impl->warningStream();
 }
 
-std::ostringstream & Logger::error()
+std::ostringstream & SimpleLogger::error()
 {
-    return m_impl->error();
+    return m_impl->errorStream();
 }
 
-std::ostringstream & Logger::fatal()
+std::ostringstream & SimpleLogger::fatal()
 {
-    return m_impl->fatal();
+    return m_impl->fatalStream();
 }
 
-std::string Logger::version()
+std::string SimpleLogger::version()
 {
-    return "1.4.0";
+    return "2.0.0";
 }
 
-Logger::~Logger() = default;
+SimpleLogger::~SimpleLogger() = default;
 
 } // juzzlin
